requirements.txt
python_log_indenter
numpy
commentjson
PyYAML
pydantic
Jinja2
jq
fastapi
uvicorn[standard]
asciinet @ git+https://github.com/cosminbasca/asciinet.git#subdirectory=pyasciinet
py-dagviz @ git+https://github.com/ctongfei/py-dagviz.git